new DOMhelpers & FunctionalUtils services
