+++ projects/ui-framework/src/lib/services/truncate-tooltip/truncate-tooltip.interface.ts
@@ -1,12 +1,3 @@
 export interface Styles {
   [key: string]: string | number;
 }
-
-export interface TextProps {
-  [key: string]: number;
-}
-
-export interface NotEmptyChildren {
-  total: number;
-  firstIndex: number | null;
-}
